requirements.txt
python-dotenv ## This is specifically used to call our environment variables
pandas
numpy
pymongo ## Specifically will be used for data ingestion
certifi
scikit-learn
dill  ## Specifically used for pickling our files
pyaml
mlflow
dagshub
fastapi
uvicorn

-e .  ## This is responsible for triggering the setup.py file. ---> Automatically building all the packages
# this -e . --> u can actually add at the last, when ur project is ready and u want to create the final package at that time u can finally put this.
## this is usually used when u want to deploy the packages into the cloud or deploy as a pypi package.
## Add -e . at the last.